+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Auto apply changes in priority, project, context and key-value dialogs #65
 - Auto apply projects and contexts tags if a new todo is created within the filter page #44
+- Context and project tags will no longer change to lower case #64
 
 ## [0.10.1] - 2024-09-22